+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- `--format=json` support for `diff`, `tree`, `stale`, and `why` commands
 - Ignore go.sum (checksums only), treat go.mod as lockfile
 
 ## [0.6.1] - 2026-01-05